Profile Page redirection based on user access.
If someone goes to view a user profile, then for an unauthorized users to set a referrer’s cookie (as with url /invite/) and show 403. For an authorized users when viewing a profile, redirect to a profile by nickname if it is, and show the same information in the account in the Profile section, these are ID, Name and Username. Example for public profile url: /u/12