fixes for part 1 (#256)
* Fixes for support groups "/support-groups"

Update to Fixes for support groups "/support-groups" #254

* Fixes for support groups added auto start command

When the user creates a support group, you need to automatically create a "/start" command, in all languages that are connected to the support group, with the message "Welcome!". This command will automatically enable the value "Is default".

* Fixes for part 1.

When the user creates a support group, you need to automatically create a "/start" command, in all languages that are connected to the support group, with the message "Welcome!". This command will automatically enable the value "Is default".